clippy fix
